refactor(meta): create a `png/error.rs`
